Move plugin test files to plugins folder
